fix(client): tighten auth + onboarding for beta launch
Polish pass driven by visual QA of the auth + onboarding flow.

Login screen
- Add prominent "Echo is in beta" callout (BetaBanner) above the
  username field. Carries icon, short copy, and a tappable Learn-more
  chevron pointing at echo-messenger.us/beta. Moved here from the
  welcome step in the wizard so new signups see it before they sign
  up and returning testers see it on every login.
- Drop the "New here?" divider/subtitle and the "Trouble signing in?"
  mailto button.  The Create-an-account TextButton is now the only
  affordance below "Forgot password?".

Onboarding wizard
- Welcome step: drop the "About 90 seconds…" sidebar subtitle, the
  "Set up your profile." sub-heading, and the old beta banner.
- Theme picker: reduce the catalog from six near-duplicate palettes
  to a curated three — System, Paper, Ember.  Other themes still
  ship and stay accessible under Settings → Appearance.
- Theme picker: replace the single-color swatch with the same rich
  mini chat-mockup the Settings picker uses (sidebar strip +
  received/sent bubble preview, split dark/light mock for System).
  The thumbnail widgets and palette presets moved from
  appearance_section.dart into a shared widgets/theme_thumbnail.dart.
- Footer / sidebar: drop the "Step N of 5" line.  The sidebar step
  list is now click-back-able — completed steps are buttons that
  jump back; future steps stay un-clickable so a rushed user can't
  skip required setup.
- Encryption step: simplify to a short privacy reassurance ("Private
  by default. Your messages are end-to-end encrypted.") and lose the
  Signal-Protocol jargon + Safety-Number paragraph plus the empty
  illustration box.
